runner/pyproject.toml
[project]
name = "arcadeops-control-tower-runner"
version = "0.1.0"
description = "FastAPI runner for the ArcadeOps Control Tower hackathon demo."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "python-dotenv>=1.0",
  "google-genai>=0.3",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
